TOTS-62 Refactored the generate random accordion item.
